+++ Lifestream/Services/TerritoryWatcher.cs
@@ -26,7 +26,7 @@ public static class TerritoryWatcher
 
     public static bool IsDataReliable() => LastHousingOutdoorTerritory != 0;
 
-    private static void ClientState_TerritoryChanged(ushort obj)
+    private static void ClientState_TerritoryChanged(uint obj)
     {
         if(Utils.IsTerritoryResidentialDistrict(obj))
         {
@@ -37,7 +37,7 @@ public static class TerritoryWatcher
 
     public static ushort GetRealTerritoryType()
     {
-        if(Svc.ClientState.TerritoryType.EqualsAny<ushort>(Houses.Private_Cottage_Empyreum, Houses.Private_Cottage_Mist, Houses.Private_Cottage_Shirogane, Houses.Private_Cottage_The_Goblet, Houses.Private_Cottage_The_Lavender_Beds, 1249))
+        if(Svc.ClientState.TerritoryType.EqualsAny<uint>(Houses.Private_Cottage_Empyreum, Houses.Private_Cottage_Mist, Houses.Private_Cottage_Shirogane, Houses.Private_Cottage_The_Goblet, Houses.Private_Cottage_The_Lavender_Beds, 1249))
         {
             return LastHousingOutdoorTerritory switch
             {
@@ -46,10 +46,10 @@ public static class TerritoryWatcher
                 ResidentalAreas.The_Goblet => Houses.Private_Cottage_The_Goblet,
                 ResidentalAreas.Shirogane => Houses.Private_Cottage_Shirogane,
                 ResidentalAreas.Empyreum => Houses.Private_Cottage_Empyreum,
-                _ => Svc.ClientState.TerritoryType
+                _ => (ushort)Svc.ClientState.TerritoryType
             };
         }
-        if(Svc.ClientState.TerritoryType.EqualsAny<ushort>(Houses.Private_House_Empyreum, Houses.Private_House_Mist, Houses.Private_House_Shirogane, Houses.Private_House_The_Goblet, Houses.Private_House_The_Lavender_Beds, 1250))
+        if(Svc.ClientState.TerritoryType.EqualsAny<uint>(Houses.Private_House_Empyreum, Houses.Private_House_Mist, Houses.Private_House_Shirogane, Houses.Private_House_The_Goblet, Houses.Private_House_The_Lavender_Beds, 1250))
         {
             return LastHousingOutdoorTerritory switch
             {
@@ -58,10 +58,10 @@ public static class TerritoryWatcher
                 ResidentalAreas.The_Goblet => Houses.Private_House_The_Goblet,
                 ResidentalAreas.Shirogane => Houses.Private_House_Shirogane,
                 ResidentalAreas.Empyreum => Houses.Private_House_Empyreum,
-                _ => Svc.ClientState.TerritoryType
+                _ => (ushort)Svc.ClientState.TerritoryType
             };
         }
-        if(Svc.ClientState.TerritoryType.EqualsAny<ushort>(Houses.Private_Mansion_Empyreum, Houses.Private_Mansion_Mist, Houses.Private_Mansion_Shirogane, Houses.Private_Mansion_The_Goblet, Houses.Private_Mansion_The_Lavender_Beds, 1251))
+        if(Svc.ClientState.TerritoryType.EqualsAny<uint>(Houses.Private_Mansion_Empyreum, Houses.Private_Mansion_Mist, Houses.Private_Mansion_Shirogane, Houses.Private_Mansion_The_Goblet, Houses.Private_Mansion_The_Lavender_Beds, 1251))
         {
             return LastHousingOutdoorTerritory switch
             {
@@ -70,9 +70,9 @@ public static class TerritoryWatcher
                 ResidentalAreas.The_Goblet => Houses.Private_Mansion_The_Goblet,
                 ResidentalAreas.Shirogane => Houses.Private_Mansion_Shirogane,
                 ResidentalAreas.Empyreum => Houses.Private_Mansion_Empyreum,
-                _ => Svc.ClientState.TerritoryType
+                _ => (ushort)Svc.ClientState.TerritoryType
             };
         }
-        return Svc.ClientState.TerritoryType;
+        return (ushort)Svc.ClientState.TerritoryType;
     }
 }
